backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.15.0
pillow==10.1.0
python-multipart==0.0.6
numpy==1.24.3
wikipedia-api
supabase
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
psycopg2-binary
pydantic 